login, owner

+++ StudyReminder/Reminder.API/Controllers/LessonController.cs
@@ -2,16 +2,20 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Reminder.API.Contracts;
 using Reminder.API.Contracts.Requests;
 using Reminder.API.Entities;
+using Reminder.API.Extensions;
 using Reminder.API.Services;
 
 namespace Reminder.API.Controllers
 {
     [ApiController]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public class LessonController : ControllerBase
     {
         private readonly ILessonService _LessonService;
@@ -23,7 +27,7 @@ namespace Reminder.API.Controllers
         [HttpGet(ApiRoutes.Lesson.GetAll)]
         public async Task<IActionResult> GetAll()
         {
-            var lesson = await _LessonService.GetLessons();
+            var lesson = await _LessonService.GetLessons(HttpContext.GetUserId());
 
             return Ok(lesson);
         }
@@ -31,6 +35,12 @@ namespace Reminder.API.Controllers
         [HttpGet(ApiRoutes.Lesson.Get)]
         public async Task<IActionResult> Get([FromRoute] Guid lessonId)
         {
+            var userOwnLesson = await _LessonService.UserOwnsLesson(lessonId, HttpContext.GetUserId());
+            if (!userOwnLesson)
+            {
+                return BadRequest(new { error = "You do not own this lesson" });
+            }
+
             var lesson = await _LessonService.GetLessonById(lessonId);
 
             if (lesson == null)
@@ -49,7 +59,9 @@ namespace Reminder.API.Controllers
                 Code = lessonRequest.Code,
                 Name = lessonRequest.Name,
                 Day = lessonRequest.Day,
-                Time = lessonRequest.Time
+                Time = lessonRequest.Time,
+                UserId = HttpContext.GetUserId()
+                
             };
 
             await _LessonService.CreateLesson(lesson);
@@ -63,6 +75,12 @@ namespace Reminder.API.Controllers
         [HttpPut(ApiRoutes.Lesson.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid lessonId, [FromBody] UpdateLessonRequest lessonRequest)
         {
+            var userOwnLesson = await _LessonService.UserOwnsLesson(lessonId, HttpContext.GetUserId());
+            if (!userOwnLesson)
+            {
+                return BadRequest(new { error = "You do not own this lesson" });
+            }
+
             var lesson = new Lesson
             {
                 Id = lessonId.ToString(),
@@ -84,6 +102,12 @@ namespace Reminder.API.Controllers
         [HttpDelete(ApiRoutes.Lesson.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid lessonId)
         {
+            var userOwnLesson = await _LessonService.UserOwnsLesson(lessonId, HttpContext.GetUserId());
+            if (!userOwnLesson)
+            {
+                return BadRequest(new { error = "You do not own this lesson" });
+            }
+
             var deleted = await _LessonService.DeleteLesson(lessonId);
 
             if (deleted)
@@ -91,5 +115,9 @@ namespace Reminder.API.Controllers
 
             return NotFound();
         }
+
+        
     }
+
+    
 }
